Adds Filestash blueprint (#218)
* Adds Filestash blueprint

Adds a new blueprint for deploying Filestash, a web-based file manager, along with a Collabora Online instance for document editing.

The blueprint includes:
- Docker Compose configuration for Filestash and Collabora Online.
- Configuration template for setting up domains and environment variables.
- An SVG logo for Filestash.

This enables users to easily deploy a complete Filestash instance with document editing capabilities.

* Renames file to docker-compose.yml

* Adds Filestash app to the meta.json

Adds Filestash, a file manager that connects storage with identity providers and authorizations, to the list of available applications.

* Config for domain handling

Simplifies domain configuration by removing the dedicated
`wopi_server` domain and adjusting the APPLICATION_URL. This
streamlines the setup and ensures correct URL handling.

Adds a comment in the docker-compose file to indicate SSL configuration.